tools: add TP-Link split U-Boot support
Newer TP-Link MediaTek devices are using a split U-Boot design with a
factory- and a firmware-uboot. The factory U-Boot contains the recovery
(HTTP and TFTP), unpacking and starting of the kernel is done in the
firmware-uboot.

All installation-mathods (Web-UI / HTTP-recovery / tftp-recovery) expect
a firmware-uboot to be in place. This makes generating images with an
integrated U-Boot mandatory. Images are pre-compiled from TP-Link GPL
tarball as they can only be compiled with their proprietary toolchain.

Firmware-images are expected to have two headers (One for the
kernel+rootfs, one for the entire image). Modify mktplinkfw to allow for
creation of the header for the entire image.

Signed-off-by: David Bauer <mail@david-bauer.net>